+++ findingaids/settings.py
@@ -209,7 +209,6 @@ except ImportError:
     pass
 
 # django_nose configurations
-
 django_nose = None
 try:
     # NOTE: errors if DATABASES is not configured (in some cases),
@@ -217,7 +216,6 @@ try:
     import django_nose
 except ImportError:
     pass
-
 
 # - only if django_nose is installed, so it is only required for development
 if django_nose is not None:
@@ -233,3 +231,28 @@ if django_nose is not None:
 # against non-test configured existdb collection
 else:
     TEST_RUNNER = 'eulexistdb.testutil.ExistDBTextTestSuiteRunner'
+
+# enable django-debug-toolbar when available & in debug/dev modes
+if DEBUG or DEV_ENV:
+    try:
+        import debug_toolbar
+        INSTALLED_APPS.append('debug_toolbar')
+    except ImportError:
+        pass
+
+# configure: default toolbars + existdb query panel
+DEBUG_TOOLBAR_PANELS = [
+    'debug_toolbar.panels.versions.VersionsPanel',
+    'debug_toolbar.panels.timer.TimerPanel',
+    'debug_toolbar.panels.settings.SettingsPanel',
+    'debug_toolbar.panels.headers.HeadersPanel',
+    'debug_toolbar.panels.request.RequestPanel',
+    'eulexistdb.debug_panel.ExistDBPanel',
+    'debug_toolbar.panels.sql.SQLPanel',
+    'debug_toolbar.panels.staticfiles.StaticFilesPanel',
+    'debug_toolbar.panels.templates.TemplatesPanel',
+    'debug_toolbar.panels.cache.CachePanel',
+    'debug_toolbar.panels.signals.SignalsPanel',
+    'debug_toolbar.panels.logging.LoggingPanel',
+    'debug_toolbar.panels.redirects.RedirectsPanel',
+]